リソースコントローラー用のルート設定を追加
- 競合する部分をコメントアウト、もしくは退避
