+++ routes/web.php
@@ -16,6 +16,8 @@ use App\Http\Controllers\PostController;
 |
 */
 
+Route::resource('post', PostController::class);
+
 Route::get('/', function () {
     return view('welcome');
 })
@@ -24,7 +26,7 @@ Route::get('/', function () {
 
 Route::get('/dashboard', function () {
     return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+})->name('dashboard');
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
@@ -32,21 +34,21 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-Route::get('/test', [TestController::class, 'test'])->name('test');
-Route::post('post', [PostController::class, 'store'])->name('post.store');
-// Route::middleware(['auth', 'can:admin'])->group(function() {
-    Route::get('post', [PostController::class, 'index'])->name('post.index');
-    Route::get('post/create', [PostController::class, 'create'])->name('create');
-// });
+// Route::get('/test', [TestController::class, 'test'])->name('test');
+// Route::post('post', [PostController::class, 'store'])->name('post.store');
+// // Route::middleware(['auth', 'can:admin'])->group(function() {
+//     Route::get('post', [PostController::class, 'index'])->name('post.index');
+//     Route::get('post/create', [PostController::class, 'create'])->name('create');
+// // });
 
-// CRUDのルーティングを定義
-// postの個別ページ
-Route::get('post/show/{post}', [PostController::class, 'show'])->name('post.show');
-// 編集ページ
-Route::get('post/{post}/edit', [PostController::class, 'edit'])->name('post.edit');
-// 更新処理 patchの代わりにputでも問題ない
-Route::patch('post/{post}', [PostController::class, 'update'])->name('post.update');
-// 削除処理
-Route::delete('post/{post}', [PostController::class, 'destroy'])->name('post.destroy');
+// // CRUDのルーティングを定義
+// // postの個別ページ
+// Route::get('post/show/{post}', [PostController::class, 'show'])->name('post.show');
+// // 編集ページ
+// Route::get('post/{post}/edit', [PostController::class, 'edit'])->name('post.edit');
+// // 更新処理 patchの代わりにputでも問題ない
+// Route::patch('post/{post}', [PostController::class, 'update'])->name('post.update');
+// // 削除処理
+// Route::delete('post/{post}', [PostController::class, 'destroy'])->name('post.destroy');
 
 require __DIR__.'/auth.php';
